feat: Add functionality to retrieve route data from CN Rail API
This commit adds code to retrieve route data from the CN Rail API. It uses the httpx library to make HTTP requests and retrieves the data in JSON format. The retrieved data is then appended to a list.

The code includes a function `get_route_data` that takes an index and a base URL as parameters. It constructs the URL with the index and makes a GET request to retrieve the data. If the request is successful (status code 200), the JSON response is appended to the `data` list.

The code also includes a main block that sets the multiprocessing start method to "spawn" and iterates over a list of characters to construct the URL. For each character, it creates a list of tasks, where each task is a multiprocessing process that calls the `get_route_data` function with a different index. The tasks are then started and joined to ensure all processes complete before continuing.

Finally, the retrieved data is dumped to a JSON file named "data.json".

This commit adds the initial implementation of the route data retrieval functionality.